Moved StatisticFunction enum to separate module in services/utils
