+++ backend/src/fastapi_app/routers/timeseries/converters.py
@@ -1,7 +1,8 @@
 from typing import List, Optional, Sequence
 
-from ....services.summary_vector_statistics import StatisticFunction, VectorStatistics
+from ....services.summary_vector_statistics import VectorStatistics
 from ....services.sumo_access.summary_access import VectorMetadata
+from ....services.utils.statistic_function import StatisticFunction
 from . import schemas
