+++ backend/src/services/utils/statistic_function.py
@@ -1,0 +1,20 @@
+from enum import Enum
+from typing import Optional
+
+
+class StatisticFunction(Enum):
+    MIN = "MIN"
+    MAX = "MAX"
+    MEAN = "MEAN"
+    P10 = "P10"
+    P90 = "P90"
+    P50 = "P50"
+
+    @classmethod
+    def from_string_value(cls, value: str) -> Optional["StatisticFunction"]:
+        try:
+            return cls(value)
+        except ValueError:
+            return None
+
+
